arrang test helpers and rename conn vm methods
